+++ sdk/src/logs/logger.cc
@@ -162,6 +162,7 @@ opentelemetry::nostd::unique_ptr<opentelemetry::logs::LogRecord> Logger::CreateL
 
   auto recordable = context_->GetProcessor().MakeRecordable();
 
+  recordable->SetLogRecordLimits(context_->GetLogRecordLimits());
   recordable->SetObservedTimestamp(std::chrono::system_clock::now());
 
   StampSpanContextFromVariant(
@@ -184,6 +185,7 @@ opentelemetry::nostd::unique_ptr<opentelemetry::logs::LogRecord> Logger::CreateL
 
   auto recordable = context_->GetProcessor().MakeRecordable();
 
+  recordable->SetLogRecordLimits(context_->GetLogRecordLimits());
   recordable->SetObservedTimestamp(std::chrono::system_clock::now());
 
   StampSpanContextFromVariant(context_or_span, *recordable);